#24 fix seminars search and refactor
